Rename once.
